Massive refactoring.

+++ include/stg/admin.h
@@ -32,7 +32,7 @@ public:
     virtual const std::string & GetPassword() const = 0;
     virtual const std::string & GetLogin() const = 0;
     virtual PRIV const *        GetPriv() const = 0;
-    virtual uint16_t            GetPrivAsInt() const = 0;
+    virtual uint32_t            GetPrivAsInt() const = 0;
     virtual const ADMIN_CONF &  GetConf() const = 0;
     virtual uint32_t            GetIP() const = 0;
     virtual std::string         GetIPStr() const = 0;
